fix(compose): Show only one suggestions bar per compose window
The suggestion list is now shared by to, cc and bcc.
