+++ src/app/compose/mailrecipientinput.component.ts
@@ -17,7 +17,7 @@
 // along with Runbox 7. If not, see <https://www.gnu.org/licenses/>.
 // ---------- END RUNBOX LICENSE ----------
 
-import { Component, Input, EventEmitter, Output, OnInit, ViewChild, ElementRef, AfterViewInit } from '@angular/core';
+import { Component, Input, EventEmitter, Output, OnChanges, ViewChild, ElementRef, AfterViewInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { BehaviorSubject } from 'rxjs';
 import { MatAutocomplete } from '@angular/material/autocomplete';
@@ -25,7 +25,6 @@ import { MatChipInputEvent } from '@angular/material/chips';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ENTER } from '@angular/cdk/keycodes';
 import { debounceTime } from 'rxjs/operators';
-import { MailAddressInfo } from '../xapian/messageinfo';
 import { RecipientsService } from './recipients.service';
 import { Recipient } from './recipient';
 
@@ -37,17 +36,11 @@ const COMMA = 188;
     selector: 'mailrecipient-input',
     templateUrl: 'mailrecipientinput.component.html'
 })
-export class MailRecipientInputComponent implements OnInit, AfterViewInit {
+export class MailRecipientInputComponent implements OnChanges, AfterViewInit {
     filteredRecipients: BehaviorSubject<Recipient[]> = new BehaviorSubject([]);
 
     searchTextFormControl: FormControl = new FormControl();
     recipientsList: string[] = [];
-
-    // here we keep the suggestions we get from RecipientsService
-    suggestedRecipients: Recipient[] = [];
-    // and here we keep suggestedRecipients but filtered
-    // to not include the recipients already picked
-    filteredSuggestions: Recipient[] = [];
 
     separatorKeysCodes = [COMMA, ENTER];
 
@@ -86,14 +79,9 @@ export class MailRecipientInputComponent implements OnInit, AfterViewInit {
             }
             );
         });
-
-        recipientservice.recentlyUsed.subscribe(suggestions => {
-            this.suggestedRecipients = suggestions;
-            this.updateSuggestions();
-        });
     }
 
-    ngOnInit() {
+    ngOnChanges() {
         this.recipientsList = this.recipients ?
             this.recipients.split(',').map((recipient) => recipient.trim()) :
             [];
@@ -107,7 +95,6 @@ export class MailRecipientInputComponent implements OnInit, AfterViewInit {
 
     notifyChangeListener() {
         this.updateRecipient.emit(this.recipientsList.join(','));
-        this.updateSuggestions();
     }
 
     removeRecipient(ndx: number) {
@@ -175,14 +162,5 @@ export class MailRecipientInputComponent implements OnInit, AfterViewInit {
         }
 
         this.addedFromAutoComplete = false;
-    }
-
-    /// updates the displayed `suggestedRecipients`
-    /// making sure it doesn't contain any existing recipients
-    updateSuggestions() {
-        const keyOf = (addr: string) => MailAddressInfo.parse(addr)[0].address;
-        this.filteredSuggestions = this.suggestedRecipients.filter(
-            s => !this.recipientsList.find(r => keyOf(r) === keyOf(s.toString()))
-        );
     }
 }
